\name{dfsaneacc-package}
\alias{dfsaneacc-package}
\docType{package}
\title{
  \packageTitle{dfsaneacc}
}
\description{
  \packageDescription{dfsaneacc}
}
\details{
  This package includes the function:
  \preformatted{
    dfsaneacc function to solve large-scale nonlinear systems of
              equations using a derivative-free approach with sequential
              secant acceleration for spectral residual methods.
  }
}
\author{
  \packageAuthor{dfsaneacc}

  Maintainer: \packageMaintainer{dfsaneacc}
}
\references{
  J. Barzilai, and J. M. Borwein (1988), Two-point step size gradient
  methods, \emph{IMA J Numerical Analysis}, 8, 141-148.

  E. G. Birgin, J. M. Mart\'inez (2021), Secant acceleration of sequential
  residual methods for solving large-scale nonlinear systems of
  equations. \emph{Technical Report arXiv:2012.13251v1,arXiv}. URL:
  \url{https://arxiv.org/abs/2012.13251}.

  W. LaCruz, and M. Raydan (2003), Nonmonotone spectral methods for
  large-scale nonlinear systems, \emph{Optimization Methods and
  Software}, 18, 583-599.

  W. LaCruz, J. M. Mart\'inez, and M. Raydan (2006), Spectral residual
  method without gradient information for solving large-scale nonlinear
  systems of equations, \emph{Mathematics of Computation}, 75,
  1429-1448.

  M. Raydan (1997), Barzilai-Borwein gradient method for large-scale
  unconstrained minimization problem, \emph{SIAM J of Optimization}, 7,
  26-33.
}
\keyword{package}
